代码检查，问题修改（补充部分函数的xCore.iRet返回值）

+++ xrt.h
@@ -394,8 +394,8 @@
 	XXAPI int xrtPathIsRelW(wstr sPath, size_t iSize);
 	
 	// 获取随机不存在的路径（ 需使用 xrtFree 释放内存 ）
-	XXAPI ustr xrtPathRandom(ustr sHead, size_t iHeadSize, ustr sFoot, size_t iFootSize, int iLen);
-	XXAPI wstr xrtPathRandomW(wstr sHead, size_t iHeadSize, wstr sFoot, size_t iFootSize, int iLen);
+	XXAPI ustr xrtPathRandom(ustr sHead, size_t iHeadSize, ustr sFoot, size_t iFootSize, size_t iLen);
+	XXAPI wstr xrtPathRandomW(wstr sHead, size_t iHeadSize, wstr sFoot, size_t iFootSize, size_t iLen);
 	
 	// 拼接路径（ 需要使用 xrtFree 释放内存 ）
 	XXAPI ustr xrtPathJoin(uint iCount, ...);
